[64 BITS] * BUG: Fixed some declarations in the gb.net.curl component. * NEW: Pointer[] is a new class that represents an array of Pointer values. * BUG: All gb.qt, gb.gtk and gb.desktop methods that deal with X11 Window   identifiers now use Pointer and Pointer[] datatypes instead of Integer   and Integer[]. * BUG: The Embedder example was updated according to the change above.
git-svn-id: svn://localhost/gambas/branches/2.0@1032 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ gb.gtk/src/gcontrol.h
@@ -32,7 +32,7 @@ public:
 	virtual bool enabled();
 	bool expand();
 	bool ignore();
-	virtual int handle();
+	virtual long handle();
 	virtual int height();
 	virtual int left();
 	int x() { return left(); }
